Fix migration number collisions
This commit fixes a bug where if there were multiple migrations with the
same migration number, only one of the migrations would run.

Instead, we now return an error if we detect multiple migrations with
the same number.

So, a directory that looks like this will now lead to an error:

```
0001.first-migration.up.sql
0001.first-migration.down.sql
0002.janes-migration.up.sql
0002.janes-migration.down.sql
0002.robs-migration.up.sql
0002.robs-migration.down.sql
```

Closed https://github.com/Keats/dbmigrate/issues/47